Open web page before launching server
Before, the web page would only launch after you killed the server with Control-C, because the `sbt run` step runs forever. Now, I open the web page beforehand, and also print a warning that it won’t work until the server is actually running.

I don’t know of an easy to way to make the `sbt run` task notice when the server is ready, and at that time, trigger some function in the shell script to open the web page.